Move tests into their own file
