+++ Tests/ObservedObjectTests/PropertyPublishingTests.swift
@@ -155,109 +155,12 @@ class PropertyPublishingTests: XCTestCase {
         cancellable.cancel()
     }
     
-    func testPublishingTwoProperties_InitialValue() {
-        class ContainerWithTwoProperties<A, B>: ObservedObject {
-            @Observed var first: A?
-            @Observed var second: B?
-        }
-        
-        let container = ContainerWithTwoProperties<Int, Int>()
-        container.first = 1
-        container.second = 2
-        
-        var observed: (Int?, Int?)?
-        let cancellable = container
-            .publisher(for: \.first, \.second)
-            .sink { (first, second) in
-                observed = (first, second)
-            }
-        
-        XCTAssertEqual(1, observed?.0)
-        XCTAssertEqual(2, observed?.1)
-        
-        cancellable.cancel()
-    }
-    
-    func testPublishingTwoProperties_FirstValueChanges() {
-        class ContainerWithTwoProperties<A, B>: ObservedObject {
-            @Observed var first: A?
-            @Observed var second: B?
-        }
-        
-        let container = ContainerWithTwoProperties<Int, Int>()
-        container.first = 1
-        container.second = 2
-        
-        var observed: (Int?, Int?)?
-        let cancellable = container
-            .publisher(for: \.first, \.second)
-            .sink { (first, second) in
-                observed = (first, second)
-            }
-        
-        container.first = 10
-        
-        XCTAssertEqual(10, observed?.0)
-        XCTAssertEqual(2, observed?.1)
-        
-        cancellable.cancel()
-    }
-    
-    func testPublishingTwoProperties_SecondValueChanges() {
-        class ContainerWithTwoProperties<A, B>: ObservedObject {
-            @Observed var first: A?
-            @Observed var second: B?
-        }
-        
-        let container = ContainerWithTwoProperties<Int, Int>()
-        container.first = 1
-        container.second = 2
-        
-        var observed: (Int?, Int?)?
-        let cancellable = container
-            .publisher(for: \.first, \.second)
-            .sink { (first, second) in
-                observed = (first, second)
-            }
-        
-        container.second = 20
-        
-        XCTAssertEqual(1, observed?.0)
-        XCTAssertEqual(20, observed?.1)
-        
-        cancellable.cancel()
-    }
-    
     private struct NonEquatable {
         
         var value: String
         
         func equals(other: NonEquatable) -> Bool {
             self.value == other.value
-        }
-        
-    }
-    
-    private class SpySubscriber<T, U>: Combine.Subscriber where U: Error {
-                
-        typealias Input = T
-        typealias Failure = U
-        
-        private(set) var subscription: Subscription?
-        func receive(subscription: Subscription) {
-            self.subscription = subscription
-        }
-        
-        private(set) var completion: Subscribers.Completion<U>?
-        func receive(completion: Subscribers.Completion<U>) {
-            self.completion = completion
-        }
-        
-        private(set) var receiveInputs = [T]()
-        var demand: Subscribers.Demand = .unlimited
-        func receive(_ input: T) -> Subscribers.Demand {
-            receiveInputs.append(input)
-            return demand
         }
         
     }
